Code rabbit review fixes
